+++ controls-constructor/src/commonMain/kotlin/space/kscience/controls/constructor/VirtualValueState.kt
@@ -8,7 +8,6 @@ import kotlinx.coroutines.flow.flowOf
 /**
  * A [MutableValueState] that does not correspond to a physical state
  *
- * @param callback a synchronous callback that could be used without a scope
  */
 private class VirtualValueState<T>(
     initialValue: T,
@@ -33,19 +32,18 @@ private class VirtualValueState<T>(
 
 
 /**
- * A [MutableDeviceState] that does not correspond to a physical state
+ * A [MutableValueState] that does not correspond to a physical state
  *
- * @param callback a synchronous callback that could be used without a scope
  */
-public fun <T> MutableDeviceState(
+public fun <T> MutableValueState(
     initialValue: T,
 ): MutableValueState<T> = VirtualValueState(initialValue)
 
 
 /**
- * Create a [DeviceState] with constant value
+ * Create a [ValueState] with constant value
  */
-public fun <T> DeviceState(
+public fun <T> ValueState(
     value: T
 ): ValueState<T> = object : ValueState<T> {
     override val value: T get() = value